Re-inflating view in onConfigurationChange
We change the behavior of adapting to theme changes to re-inflate the
view and hook the method "onConfigurationChange" for the setting panel
to the controller. All resources are configurated in resource files.

To achieve theme change for setting panel:
1. We hide/remove the panel from screen if it is visible.
2. We inflate the view again. The contents in "inflateView" are
   extracted from the origin constructor.
3. We show the panel if it is opened before the theme changes.

We only move resource configurations to resource files and the contents
of resources remain the same.

Bug: 257272333
Test: manual - attach screenshots with the bug
Change-Id: I6fe7b800a9fabc25a192749896480446347f99d7