Add context-aware AudioTrack construction APIs.
This cl also introduces new device policy type for audio.
The POLICY_TYPE_AUDIO is used to specialize AudioManager
behavior for device-specific context - if the context
belongs to VirtualDevice instance with custom device policy
for audio, it will try to extract the device specific session
id for that device from VDM. The device specific session aid
is used to aid virtual device audio routing.

The context-aware constructor and builder will be extended
in follow-up cls to extract AttributionSource from the Context
and pass it to the native layer.

The policy will be used to specialize AudioManager behavior for
device-specific context.

Bug: 261698699
Bug: 249777386
Test: atest VirtualDeviceParamsTest AudioTrackUnitTest AudioTrackTest
Change-Id: I558b508f06e8098a8fbebaa0560557af06cdd0f7