[Unfold transition] Fully manage the vignette on the background thread
Updates onScreenTurningOn/device state callbacks
in the unfold overlay class to be executed on
the background thread.

This allows to draw the first black frame even
if SystemUI's main thread is busy at the moment,
so it will be less likely to have flicker because
of that.

Also this CL adds a separate thread to render
the vignette to avoid conflicting with the other
operations made on the shared background thread.

Bug: 262381044
Bug: 262518402
Bug: 233045200
Test: manual fold/unfolds with/without AOD enabled
Test: add artifical delay to SystemUI configChanged
 that exceeds WM keyguard drawn timeout
 => test that there is a flicker without changes
    test that there is no flicker with the changes
Test: atest com.android.keyguard.mediator.ScreenOnCoordinatorTest
Change-Id: Ic5a61839500dbbf63e9c5422bfc0085827526d92