Add CompatChange flag to disable new launching app identity access
Android U introduced new APIs that allow a launching app to share its
identity with the launched activity; the activity could then access
the identity of the app by using the previously hidden
getLaunchedFrom APIs. This commit adds a new compat change that
allows the new requirements to access the launching app's identity
(app opted-in through ActivityOptions or uid of the activity is the
same as the app) to be disabled through the compatibility framework.

Bug: 259743961
Test: atest ShareIdentityTest with compat change enabled and disabled
Change-Id: Id3647611cf0f89ccaba75b24b20ad6f616310606