Handle dream restart
When a startDream is called directly after a stopDream for the same
DreamService, the old instance should be stopped and the state
should be cleared. Otherwise, the old dream's
mStopStubbornDreamRunnable messes up the state in the dream manager
and power manager.

In this CL we stop the old dream immediately if a startDream is
called for the same dream component while the old dream is gently
waking. This ensures that the dream state is cleared.

Bug: 260205205
Test: 2x adb shell cmd dreams start-dreaming - verify the device is
      dreaming afterwards
Change-Id: Ibd261038430a38f2fa519549372ddbf19d7cb36b