Grant app visibility if launching app opts in to sharing identity
Android U introduced new APIs to allow a launching app to opt-in to
sharing its identity with a launched activity; this allows the
launched activity to perform authorization checks against the
launching app. This commit grants app visibility to allow the
launched activity to access the details and signing identity of
the launching app.

Bug: 262605364
Test: atest ShareIdentityTest
Change-Id: I2b4b2acbd127b01f231c4852704ba69e31c4aaed