Refreshing state in QS with upcoming state of KEYGUARD
Transition to AOD was glitchy because many/all parts of QS were not aware we're transitioning to keyguard already (upcoming state) even though current state of StatusBarState still might not be there.
This prevents QS becoming visible for a moment and then quickly disappearing when StatusBarState catches up with the upcoming state.

Fixes: 258219778
Test: On large screen portrait, enable AOD and from unlocked state press power button. QS tiles should not be visible at any point.
Change-Id: I0187f7bf0ee329d8f1ef7bb02bbcc5c69f14ec2a